Move reader functions
